beta v1.0

+++ app/routers/soundcloud_url.py
@@ -3,18 +3,21 @@ from aiogram import Router, F
 from yt_dlp import YoutubeDL
 
 from ..settings import DOWNLOAD_DIR
+from ..utils import cleanup_file
+from ..utils import spam_checker
 from .. import errors
 
 from uuid import uuid4
 import asyncio
+import re
 import os
 
 router = Router()
 
 
 # Constants
-SOUNDCLOUD_URL_REGEX = r"^(https?:\/\/)?(www\.)?(soundcloud\.com\/[\w\-]+\/[\w\-]+|on\.soundcloud\.com\/[\w]+)(\/?)$"
-MAX_FILESIZE = 3 * (1024 * 1024)  # 3 MB
+SOUNDCLOUD_URL_REGEX = r".*(https?:\/\/)?(www\.)?(soundcloud\.com\/[\w\-]+\/[\w\-]+|on\.soundcloud\.com\/[\w]+)(\/?).*"
+MAX_FILESIZE = 49 * (1024 * 1024)
 AUDIO_FORMAT = 'm4a'
 
 
@@ -23,27 +26,38 @@ async def soundcloud_url(message: Message) -> None:
     """
     Handles messages containing SoundCloud URLs, downloads the audio, and uploads it.
     """
-    alert = await message.answer("Downloading (soundcloud)...")
+    link = re.findall(r'https?://[^\s]+', message.text)[0]
+    alert = await message.answer("🔎 دریافت اطلاعات از ساوندکلاود")
     audio_path = None
+
+    seconds = spam_checker(message.from_user.id)
+    if seconds:
+        await message.reply(f"برای جلوگیری از اسپم لطفا {seconds} ثانیه دیگر تلاش کنید")
+        return
 
     try:
         # Download the audio file
-        audio_path, info = await download_audio(message.text)
+        audio_path, info = await download_audio(link)
         audio_path += ".m4a"
-        await alert.edit_text("Uploading...")
-        print(audio_path)
+        await alert.edit_text("📤 در حال آپلود...")
+
+        # Get file size in bytes
+        file_size_bytes = os.path.getsize(audio_path)
+        file_size_mb = file_size_bytes // (1024 * 1024)
+        if file_size_mb > 49:
+            raise errors.OversizeFile()
 
         # Send the audio file to the user
         await send_audio(message, audio_path, info)
         await alert.delete()
 
     except errors.OversizeFile:
-        await alert.edit_text(f"filesize more than {MAX_FILESIZE // (1024 * 1024)}MB")
+        await alert.edit_text("به دلیل محدودیت های تلگرام امکان آپلود فایل های بیشتر از ۵۰ مگابایت وجود ندارد")
     except asyncio.TimeoutError:
-        await alert.edit_text("Timeout error")
+        await alert.edit_text("مهلت انجام عملیات به پایان رسید، لطفا دوباره تلاش کنید")
     except Exception as e:
         print(f"Unexpected error: {e}")
-        await alert.edit_text("An unexpected error occurred.")
+        await alert.edit_text("یک خطای غیر منتظره رخ داد، لطفا بعدا تلاش کنید")
     finally:
         if audio_path:
             cleanup_file(audio_path)
@@ -52,10 +66,6 @@ async def soundcloud_url(message: Message) -> None:
 async def download_audio(url: str) -> tuple[str, dict]:
     """
     Downloads the audio from the given SoundCloud URL and returns the file path and metadata.
-
-    :param url: SoundCloud URL to download.
-    :return: Tuple containing the file path and metadata.
-    :raises ValueError: If file exceeds maximum allowed size or download fails.
     """
     outtmpl = (DOWNLOAD_DIR / f'{str(uuid4())}').as_posix()
     ydl_opts = {
@@ -84,7 +94,7 @@ async def download_audio(url: str) -> tuple[str, dict]:
             return outtmpl, info
 
     try:
-        return await asyncio.wait_for(asyncio.to_thread(inner), timeout=60)
+        return await asyncio.wait_for(asyncio.to_thread(inner), timeout=90)
     except Exception as e:
         raise e
 
@@ -92,10 +102,6 @@ async def download_audio(url: str) -> tuple[str, dict]:
 async def send_audio(message: Message, audio_path: str, info: dict) -> None:
     """
     Sends the downloaded audio file to the user.
-
-    :param message: Incoming message object.
-    :param audio_path: Path to the downloaded audio file.
-    :param info: Metadata of the downloaded file.
     """
     title = info.get("title", "Unknown")
     
@@ -119,19 +125,10 @@ async def send_audio(message: Message, audio_path: str, info: dict) -> None:
     except (ValueError, TypeError):
         duration = None
 
+    me = await message.bot.get_me()
     await message.answer_audio(
         audio=FSInputFile(audio_path, filename=title),
         duration=duration,
         thumbnail=URLInputFile(thumbnail) if thumbnail else None,
+        caption=f"@{me.username}",
     )
-
-
-def cleanup_file(file_path: str) -> None:
-    """
-    Deletes the file from the filesystem.
-    """
-    if file_path and os.path.exists(file_path):
-        try:
-            os.remove(file_path)
-        except Exception as e:
-            print(f"Error cleaning up file {file_path}: {e}")
